Reorder User notification in correct namespaces
