+++ app/Http/Controllers/TableController.php
@@ -130,7 +130,8 @@ class TableController extends Controller
 
         $discordNotificationData = $this->discordNotificationData::make($table->game, $table, $table->day);
 
-        $discordNotification = ($this->notificationFactory)('user-subscription', $discordNotificationData);
+        $discordNotification = ($this->notificationFactory)(entity: 'user', type: 'user-subscription',
+            discordNotificationData: $discordNotificationData);
         $discordNotification->handle();
 
         return redirect()->back();
@@ -142,7 +143,8 @@ class TableController extends Controller
 
         $discordNotificationData = $this->discordNotificationData::make($table->game, $table, $table->day);
 
-        $discordNotification = ($this->notificationFactory)('user-unsubscription', $discordNotificationData);
+        $discordNotification = ($this->notificationFactory)(entity: 'user', type: 'user-unsubscription',
+            discordNotificationData: $discordNotificationData);
         $discordNotification->handle();
 
         return redirect()->back();
