+++ app/Notifications/Discord/User/UserUnsubscriptionNotification.php
@@ -1,23 +1,24 @@
 <?php
 
-namespace App\Notifications\Discord;
+namespace App\Notifications\Discord\User;
 
 use App\Actions\Discord\SendDiscordNotificationAction;
 use App\Enums\EmbedColor;
 use App\Enums\EmbedMessageContent;
+use App\Notifications\Discord\DiscordNotification;
 use App\Notifications\Discord\Strategies\CreateMessageIntoThread;
 use Illuminate\Support\Facades\Auth;
 
-class UserSubscriptionNotification extends DiscordNotification
+class UserUnsubscriptionNotification extends DiscordNotification
 {
     public function buildMessage(): self
     {
         $this->message = [
-            'content' => EmbedMessageContent::SUBSCRIBED->value,
+            'content' => EmbedMessageContent::UNSUBSCRIBED->value,
             'embeds' => [
                 [
-                    'title' => Auth::user()->name.' s\'est inscrit à la table de '.$this->discordNotificationData->game->name,
-                    'color' => EmbedColor::CREATED->value,
+                    'title' => Auth::user()->name.' s\'est désinscrit de la table de '.$this->discordNotificationData->game->name,
+                    'color' => EmbedColor::DELETED->value,
                     'fields' => [
                         [
                             'name' => 'Date',
